jobapi/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cloudinary==1.39.0
cryptography==42.0.8
distro==1.9.0
Django==4.2.9
django-classy-tags==4.1.0
django-cleanup==8.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-debug-toolbar==4.4.6
django-mathfilters==1.0.0
django-taggit==4.0.0
django-taggit-templatetags2==1.6.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.2
Faker==19.6.0
google-auth==2.34.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
ndg-httpsclient==0.5.1
pdfminer.six==20231228
pdfplumber==0.11.2
Pillow==10.0.0
psycopg2-binary==2.9.7
pusher==3.3.2
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
rsa==4.9
setuptools==69.0.3
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wheel==0.42.0
whitenoise==6.5.0
beautifulsoup4